Generalize AddPaymentIcon component
