+++ src/components/generic/ModalIconButton/AddPaymentIcon/AddPaymentIcon.spec.ts
@@ -1,0 +1,3 @@
+describe('AddPaymentIcon', () => {
+  test.todo('Missing tests');
+});
